test: rename test_data for combines tests
